Adding slope to subgrid orography, organizing the script, creating new file eocene_functions
